+++ peppy/__init__.py
@@ -14,12 +14,13 @@ from sys import stdout
 from ._version import __version__
 from .attribute_dict import AttributeDict
 from .const import *
+from .exceptions import PeppyError
 from .project import Project, ProjectContext
 from .sample import Sample, Subsample
 
 
 __classes__ = ["AttributeDict", "Project", "Sample"]
-__all__ = __classes__
+__all__ = __classes__ + ["PeppyError"]
 
 
 LOGGING_LEVEL = "INFO"